+++ R/class_workspace.R
@@ -43,7 +43,7 @@ workspace_load_packages <- function(workspace) {
 }
 
 workspace_set_seed <- function(workspace) {
-  set.seed(workspace$target$command$seed)
+  tar_seed_set(workspace$target$command$seed)
 }
 
 workspace_validate <- function(workspace) {