+++ R/tar_seed.R
@@ -1,26 +1,13 @@
-#' @title Get the random number generator seed of the target currently running.
+#' @title Deprecated: get the seed of the current target.
 #' @export
-#' @family utilities
-#' @description Get the random number generator seed
-#'   of the target currently running.
-#' @details A target's random number generator seed
-#'   is a deterministic function of its name. In this way,
-#'   each target runs with a reproducible seed so someone else
-#'   running the same pipeline should get the same results,
-#'   and no two targets in the same pipeline share the same seed.
-#'   (Even dynamic branches have different names and thus different seeds.)
-#'   You can retrieve the seed of a completed target
-#'   with `tar_meta(your_target, seed)`
-#'   and run `set.seed()` on the result to locally
-#'   recreate the target's initial RNG state.
+#' @keywords internal
+#' @description Deprecated on 2023-10-12 (version 1.3.2.9001).
+#'   Use [tar_seed_get()] instead.
 #' @return Integer of length 1. If invoked inside a `targets` pipeline,
 #'   the return value is the seed of the target currently running,
 #'   which is a deterministic function of the target name. Otherwise,
 #'   the return value is `default`.
-#' @param default Integer, value to return if `tar_seed()`
-#'   is called on its own outside a `targets` pipeline.
-#'   Having a default lets users run things without [tar_make()],
-#'   which helps peel back layers of code and troubleshoot bugs.
+#' @inheritParams tar_seed_get
 #' @examples
 #' tar_seed()
 #' tar_seed(default = 123L)
@@ -32,12 +19,9 @@
 #' })
 #' }
 tar_seed <- function(default = 1L) {
-  default <- as.integer(default)
-  tar_assert_int(default)
-  tar_assert_scalar(default)
-  if_any(
-    !is.null(tar_runtime$target),
-    tar_runtime$target$command$seed,
-    as.integer(default)
+  tar_warn_deprecate(
+    "tar_seed() was deprecated on 2023-10-12 (version 1.3.2.9001). ",
+    "Use tar_seed_get() instead."
   )
+  tar_seed_get(default = default)
 }