+++ R/utils_digest.R
@@ -5,11 +5,11 @@ vdigest64 <- digest::getVDigest(algo = "xxhash64")
 vdigest64_file <- digest::getVDigest(algo = "xxhash64", errormode = "warn")
 
 digest_chr32 <- function(object, ...) {
-  vdigest32(object, serialize = FALSE, ...)
+  vdigest32(object, serialize = FALSE, file = FALSE, seed = 0L, ...)
 }
 
 digest_chr64 <- function(object, ...) {
-  vdigest64(object, serialize = FALSE, ...)
+  vdigest64(object, serialize = FALSE, file = FALSE, seed = 0L, ...)
 }
 
 digest_file64 <- function(object, ...) {
@@ -18,6 +18,7 @@ digest_file64 <- function(object, ...) {
     FUN = vdigest64_file,
     serialize = FALSE,
     file = TRUE,
+    seed = 0L,
     ...,
     FUN.VALUE = character(1L),
     USE.NAMES = FALSE
@@ -25,19 +26,24 @@ digest_file64 <- function(object, ...) {
 }
 
 digest_obj32 <- function(object, ...) {
-  vdigest32(list(object), serialize = TRUE, serializeVersion = 3L, ...)
+  vdigest32(
+    object = list(object),
+    serialize = TRUE,
+    serializeVersion = 3L,
+    file = FALSE,
+    seed = 0L,
+    ...
+  )
 }
 
 digest_obj64 <- function(object, ...) {
-  vdigest64(list(object), serialize = TRUE, serializeVersion = 3L, ...)
-}
-
-produce_seed <- function(scalar) {
-  seed <- tar_options$get_seed()
-  if_any(
-    anyNA(seed),
-    NA_integer_,
-    digest::digest2int(as.character(scalar), seed = seed)
+  vdigest64(
+    object = list(object),
+    serialize = TRUE,
+    serializeVersion = 3L,
+    file = FALSE,
+    seed = 0L,
+    ...
   )
 }